backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Berget1411",email = "107999361+Berget1411@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "anthropic (>=0.59.0,<0.60.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "voyageai (>=0.2.0,<0.4.0)",
    "pymupdf (>=1.23.0,<2.0.0)",
    "python-multipart (>=0.0.6,<0.1.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "pydantic (>=2.0.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
